Fix links containing just SITEURL for local dev
When the output files are generated locally the SITEURL might be set to
an empty string, which means links containing just `{{ SITEURL }}` will
cause `href=""`. In these cases the link will be to the "current" page,
not the home page as was desired.

Including `/` after the `{{ SITEURL }}` should prevent these empty
`href` string from occurring.